pyproject.toml
[project]
name = "library-management-system"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Manisha", email = "manishabhandari841@gmail.com" }
]
dependencies = [
    "fastapi[all]>=0.111.0",
    "edgedb>=1.9.0",
    "edgedb-pydantic-codegen @ git+https://github.com/dwkrghm/edgedb-pydantic-codegen",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/library_management_system"]
